Generate floats instead of doubles.
We generate floats to be sure that only one call to generator() is
enough to generate a number. Otherwise we get different results for GCC
and Visual C++, because if we generate doubles, GCC (correctly) uses two
calls to generator(), while Visual C++ uses only one, thus generating
different sequence.