Add viterbi_beam_size and viterbi_stage_pruning options.
The viterbi_stage_pruning was previously hardcoded to
Constants::prunning_constant, which is 6. This value is the default now,
but it can be modified.

Both options can be specified in the configuration file or on the
command line.

Also, for consistency decoding_order option of korektor was renamed to
viterbi_order.